requirements.txt
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.10.0
easyocr==1.7.1
editdistance==0.6.2
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gensim==4.3.2
idna==2.10
imageio==2.33.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
psutil==5.9.7
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.14
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
supervision==0.17.1
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.230
urllib3==2.1.0
zipp==3.17.0
